+++ docs/index.html
@@ -10,7 +10,10 @@
     <title>Noticioso</title>
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-    <link href="https://fonts.googleapis.com/css2?family=Alegreya:ital,wght@0,400;0,500;0,700;1,400;1,500;1,700&family=JetBrains+Mono:wght@400;500;700&display=swap" rel="stylesheet" />
+    <link
+      href="https://fonts.googleapis.com/css2?family=Alegreya:ital,wght@0,400;0,500;0,700;1,400;1,500;1,700&family=JetBrains+Mono:wght@400;500;700&display=swap"
+      rel="stylesheet"
+    />
     <style>
       :root {
         --ink: #1a1a1a;
@@ -252,11 +255,11 @@
       <img src="icon.png" alt="Noticioso icon" />
       <h1>Noticioso</h1>
       <p class="dateline">
-        Android &nbsp;·&nbsp; RSS Reader &nbsp;·&nbsp; Open Source
+        Mobile &nbsp;·&nbsp; RSS Reader &nbsp;·&nbsp; Open Source
       </p>
       <p class="lede">
-        A personal RSS feed reader built for e-ink Android devices. No
-        algorithm, no ads, no noise — just the articles you chose to follow.
+        A personal RSS reader with full article extraction, built for e-ink
+        devices. No algorithm, no noise, just the articles you chose to follow.
       </p>
       <a
         class="download-btn"
@@ -286,16 +289,12 @@
             pages, swipe up/down to navigate between screens.
           </li>
           <li>
-            <strong>Drag-to-reorder feeds</strong> — manage your feed list with
-            drag-and-drop.
+            <strong>Share to read</strong> — share any URL from your browser
+            directly into Noticioso to read it as a clean article.
           </li>
           <li>
             <strong>Import / Export</strong> — back up and restore your feed
             list as a JSON file via the native share sheet and document picker.
-          </li>
-          <li>
-            <strong>Share to read</strong> — share any URL from your browser
-            directly into Noticioso to read it as a clean article.
           </li>
         </ul>
       </section>
